--- .idea/.gitignore
@@ -1,3 +1,0 @@
-# デフォルトの無視対象ファイル
-/shelf/
-/workspace.xml
