--- Student.Management/.gitignore
@@ -1,19 +1,0 @@
-.gradle
-8.8
-buildOutputCleanup
-vcs-1
-file-system.probe
-.idea
-.gitignore
-compiler.xml
-gradle.xml
-jarRepositories.xml
-misc.xml
-uiDesigner.xml
-vcs.xml
-workspace.xml
-build
-classes
-generated
-resources
-tmp
