--- Student.Management/.gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Sun Aug 18 21:51:05 JST 2024
-gradle.version=8.8
