--- Student.Management/src/main/java/raisetech/Student/Management/Application.java
@@ -1,15 +1,0 @@
-package raisetech.Student.Management;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-
-public class Application {
-
-
-
-	public static void main(String[] args) {
-		SpringApplication.run(Application.class, args);
-	}
-}
