--- Student.Management/src/main/java/raisetech/Student/Management/Servise/StudentService.java
@@ -1,28 +1,0 @@
-package raisetech.Student.Management.Servise;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import raisetech.Student.Management.repository.StudentRepository;
-import raisetech.Student.Management.data.Student;
-import raisetech.Student.Management.data.StudentsCourses;
-
-import java.util.List;
-
-@Service
-
-public class StudentService {
-    private StudentRepository repository;
-
-    @Autowired
-    public StudentService(StudentRepository repository) {
-        this.repository = repository;
-    }
-
-    public List<Student> searchStudentList() {
-        return repository.search();
-    }
-
-    public List<StudentsCourses> searchStudentCourseList() {
-        return repository.searchStudentCourses();
-    }
-}
