--- Student.Management/src/main/java/raisetech/Student/Management/data/StudentsCourses.java
@@ -1,15 +1,0 @@
-package raisetech.Student.Management.data;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class StudentsCourses {
-
-    private String StudentId;
-    private String CourseId;
-    private String CourseName;
-    private String CourseDate;
-
-}
