--- Student.Management/src/main/resources/application.properties
@@ -1,8 +1,0 @@
-spring.application.name=StudentManagement
-
-spring.datasource.url=jdbc:mysql://localhost:3308/StudentManagement
-spring.datasource.username=root
-spring.datasource.password=Neko00814
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-mybatis.configuration.map-underscore-to-camel-case=true
-
