+++ src/main/java/raisetech/student/management/Application.java
@@ -1,0 +1,15 @@
+package raisetech.student.management;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+
+public class Application {
+
+
+
+	public static void main(String[] args) {
+		SpringApplication.run(Application.class, args);
+	}
+}
