+++ src/main/java/raisetech/student/management/controller/StudentController.java
@@ -1,0 +1,61 @@
+package raisetech.student.management.controller;
+import org.springframework.ui.Model;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
+import raisetech.student.management.controller.converter.StudentConverter;
+import raisetech.student.management.data.Student;
+import raisetech.student.management.data.StudentsCourses;
+import raisetech.student.management.domain.StudentDetail;
+import raisetech.student.management.service.StudentService;
+
+import java.util.List;
+
+@Controller
+public class StudentController {
+
+    private StudentService service;
+    private StudentConverter converter;
+
+    @Autowired
+    public StudentController(StudentService service, StudentConverter converter) {
+        this.service = service;
+        this.converter = converter;
+    }
+
+    @GetMapping("/studentList")
+    public String getStudentList(Model model) {
+        List<Student> students = service.searchStudentList();
+        List<StudentsCourses> studentsCourses = service.searchStudentCourseList();
+
+        model.addAttribute("studentList", converter.covertStudentDetails(students, studentsCourses));
+        return "studentList";
+    }
+
+
+    @GetMapping("/students_coursesList")
+    public List<StudentsCourses> getStudentcoursesList() {
+        return service.searchStudentCourseList();
+    }
+
+    @GetMapping("/newStudent")
+    public String newStudent(Model model) {
+        model.addAttribute("studentDetail", new StudentDetail());
+        return "registerStudent";
+    }
+
+    @PostMapping("/registerStudent")
+    public String registerStudent(@ModelAttribute StudentDetail studentDetail, BindingResult result){
+        result.hasErrors();
+        // 新規受講生を登録する処理を実装する。
+        // コーズ情報も一緒に登録できるように実装する。コースは単体でも良い。
+        service.registerStudent(studentDetail);
+        return "redirect:/studentList";
+    }
+}
+
+
+
