+++ src/main/java/raisetech/student/management/controller/converter/StudentConverter.java
@@ -1,0 +1,31 @@
+package raisetech.student.management.controller.converter;
+
+import org.springframework.stereotype.Component;
+import raisetech.student.management.data.Student;
+import raisetech.student.management.data.StudentsCourses;
+import raisetech.student.management.domain.StudentDetail;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Component
+public class StudentConverter {
+
+    public List<StudentDetail> covertStudentDetails(List<Student> students, List<StudentsCourses> studentsCourses) {
+        List<StudentDetail> studentDetails = new ArrayList<>();
+        students.forEach(student -> {
+            StudentDetail studentDetail = new StudentDetail();
+            studentDetail.setStudent(student);
+
+            List<StudentsCourses> convertStudentCourses = studentsCourses.stream()
+                    .filter(studentCourse -> student.getId().equals(studentCourse.getStudentId()))
+                    .collect(Collectors.toList());
+
+            studentDetail.setStudentsCourses(convertStudentCourses);
+            studentDetails.add(studentDetail);
+        });
+        return studentDetails;
+    }
+
+}
