+++ src/main/java/raisetech/student/management/data/Student.java
@@ -1,0 +1,20 @@
+package raisetech.student.management.data;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Student {
+
+    private String id;
+    private String name;
+    private String kanaName;
+    private String nickName;
+    private String email;
+    private String area;
+    private int age;
+    private String gender;
+    private String remark;
+    private boolean isDeleted;
+}
