+++ src/main/java/raisetech/student/management/data/StudentsCourses.java
@@ -1,0 +1,15 @@
+package raisetech.student.management.data;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class StudentsCourses {
+
+    private String StudentId;
+    private String CourseId;
+    private String CourseName;
+    private String CourseDate;
+
+}
