+++ src/main/java/raisetech/student/management/service/StudentService.java
@@ -1,0 +1,35 @@
+package raisetech.student.management.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import raisetech.student.management.data.Student;
+import raisetech.student.management.data.StudentsCourses;
+import raisetech.student.management.domain.StudentDetail;
+import raisetech.student.management.repository.StudentRepository;
+
+import java.util.List;
+
+@Service
+
+public class StudentService {
+    private StudentRepository repository;
+
+    @Autowired
+    public StudentService(StudentRepository repository) {
+        this.repository = repository;
+    }
+
+    public List<Student> searchStudentList() {
+        return repository.search();
+    }
+
+    public List<StudentsCourses> searchStudentCourseList() {
+
+        return repository.searchStudentCourses();
+    }
+    @Transactional
+    public void registerStudent(StudentDetail studentDetail){
+        repository.registerStudent(studentDetail.getStudent());
+    }
+}
