+++ src/main/resources/templates/registerStudent.html
@@ -1,0 +1,55 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org" xmlns="http://www.w3.org/1999/html">
+<head>
+    <meta charset = "UTF-8">
+    <title>受講生登録</title>
+</head>
+<body>
+<h1>受講生登録</h1>
+<form th:action="@{/registerStudent}" th:object="${studentDetail}" method="post">
+    <div>
+        <label for="id">id</label>
+        <input type="text" id="id" th:field="*{student.name}" required />
+    </div>
+    <div>
+        <label for="name">名前</label>
+        <input type="text" id="name" th:field="*{student.name}" required />
+    </div>
+    <div>
+        <label for="kanaName">カナ名</label>
+        <input type="text" id="kanaName" th:field="*{student.name}" required />
+    </div>
+        <div>
+        <label for="nickname">ニックネーム</label>
+        <input type="text" id="nickname" th:field="*{student.name}" required />
+    </div>
+    <div>
+        <label for="email">メールアドレス</label>
+        <input type="text" id="email" th:field="*{student.name}" required />
+    </div>
+    <div>
+        <label for="area">地域</label>
+        <input type="text" id="area" th:field="*{student.name}" required />
+    </div>
+    <div>
+        <label for="age">年齢</label>
+        <input type="text" id="age" th:field="*{student.name}" required />
+    </div>
+     <div>
+        <label for="gender">性別</label>
+        <input type="text" id="gender" th:field="*{student.name}" required />
+    </div>
+    <div>
+        <label for="remark">備考</label>
+        <input type="text" id="remark" th:field="*{student.name}" required />
+    </div>
+    <div>
+        <label for="CourseName">コース名</label>
+        <input type="text" id="CourseName" th:field="*{student.name}" required />
+    </div>
+    <div>
+        <button type="submit">登録</button>
+    </div>
+</form>
+</body>
+</html>
