+++ src/main/resources/templates/studentList.html
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org" lang="ja">
+<head>
+    <meta charset = "UTF-8">
+    <title>受講生情報</title>
+</head>
+<body>
+<h1>受講生情報</h1>
+
+<table border="1">
+    <thead>
+    <tr>
+        <th>ID</th>
+        <th>名前</th>
+        <th>カナ名</th>
+        <th>ニックネーム</th>
+        <th>メールアドレス</th>
+        <th>地域</th>
+        <th>年齢</th>
+        <th>性別</th>
+        <th>備考</th>
+    </tr>
+    </thead>
+    <tbody>
+       <tr th:each="studentDetail : ${studentList}">
+        <td th:text="${studentDetail.student.id}">1</td>
+        <td th:text="${studentDetail.student.name}">山田太郎</td>
+        <td th:text="${studentDetail.student.kanaName}">ヤマダタロウ</td>
+        <td th:text="${studentDetail.student.nickName}">タロ</td>
+        <td th:text="${studentDetail.student.email}">taro@example.com</td>
+        <td th:text="${studentDetail.student.area}">東京</td>
+        <td th:text="${studentDetail.student.age}">20</td>
+        <td th:text="${studentDetail.student.gender}">male</td>
+        <td th:text="${studentDetail.student.remark}">特になし</td>
+    </tr>
+    </tbody>
+  </table>
+</body>
+</html>
